fix: various small fixes for GCC on Linux

+++ UDT_DLL/src/analysis_cut_by_mid_air.cpp
@@ -418,7 +418,7 @@ void udtCutByMidAirAnalyzer::ProcessSnapshotMessage(const udtSnapshotCallbackArg
 
 		const udtCutByMidAirArg& extraInfo = GetExtraInfo<udtCutByMidAirArg>();
 		const s32 udtWeapon = GetUDTWeaponFromIdMod(meanOfDeath, _protocol);
-		if(udtWeapon != (u32)-1 && !IsAllowedWeapon(udtWeapon, extraInfo.AllowedWeapons))
+		if(udtWeapon != -1 && !IsAllowedWeapon(udtWeapon, extraInfo.AllowedWeapons))
 		{
 			continue;
 		}
